Move Store interface and MemStore to utils package
