+++ auth_test.go
@@ -18,11 +18,12 @@ import (
 	"testing"
 
 	"github.com/mendersoftware/mender/client"
+	"github.com/mendersoftware/mender/utils"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNewAuthManager(t *testing.T) {
-	ms := NewMemStore()
+	ms := utils.NewMemStore()
 	cmdr := newTestOSCalls("", 0)
 	idrunner := &IdentityDataRunner{
 		cmdr: &cmdr,
@@ -59,7 +60,7 @@ func TestNewAuthManager(t *testing.T) {
 }
 
 func TestAuthManager(t *testing.T) {
-	ms := NewMemStore()
+	ms := utils.NewMemStore()
 
 	cmdr := newTestOSCalls("", 0)
 
@@ -96,7 +97,7 @@ func TestAuthManager(t *testing.T) {
 }
 
 func TestAuthManagerRequest(t *testing.T) {
-	ms := NewMemStore()
+	ms := utils.NewMemStore()
 
 	var err error
 
@@ -159,7 +160,7 @@ func TestAuthManagerRequest(t *testing.T) {
 }
 
 func TestAuthManagerResponse(t *testing.T) {
-	ms := NewMemStore()
+	ms := utils.NewMemStore()
 
 	cmdr := newTestOSCalls("mac=foobar", 0)
 	am := NewAuthManager(AuthManagerConfig{
