+++ store_test.go
@@ -15,22 +15,15 @@ package main
 
 import (
 	"bytes"
-	"errors"
 	"io"
-	"io/ioutil"
-	"os"
 
+	"github.com/mendersoftware/mender/utils"
 	"github.com/stretchr/testify/mock"
-)
-
-var (
-	errDisabled = errors.New("disabled")
-	errReadOnly = errors.New("read only")
 )
 
 type MemStoreWriter struct {
 	bytes.Buffer
-	ms   *MemStore
+	ms   *utils.MemStore
 	name string
 }
 
@@ -45,96 +38,6 @@ func (msw *MemStoreWriter) Close() error {
 
 type MemStoreData struct {
 	data []byte
-}
-
-// in-memory store for testing purposes
-type MemStore struct {
-	data     map[string]*MemStoreData
-	readonly bool
-	disable  bool
-	closeErr error
-}
-
-func (ms *MemStore) OpenRead(name string) (io.ReadCloser, error) {
-	if ms.disable {
-		return nil, errDisabled
-	}
-	v, ok := ms.data[name]
-	if ok == false {
-		return nil, os.ErrNotExist
-	}
-
-	return ioutil.NopCloser(bytes.NewBuffer(v.data)), nil
-}
-
-func (ms *MemStore) OpenWrite(name string) (WriteCloserCommitter, error) {
-	if ms.disable {
-		return nil, errDisabled
-	}
-
-	if ms.readonly {
-		return nil, errReadOnly
-	}
-
-	ms.data[name] = &MemStoreData{}
-
-	msw := &MemStoreWriter{
-		bytes.Buffer{},
-		ms,
-		name,
-	}
-	return msw, nil
-}
-
-func (ms *MemStore) ReadAll(name string) ([]byte, error) {
-	in, err := ms.OpenRead(name)
-	if err != nil {
-		return nil, err
-	}
-
-	return ioutil.ReadAll(in)
-}
-
-func (ms *MemStore) WriteAll(name string, data []byte) error {
-	out, err := ms.OpenWrite(name)
-	if err != nil {
-		return err
-	}
-
-	_, err = out.Write(data)
-	return out.Commit()
-}
-
-func (ms *MemStore) Close() error {
-	return ms.closeErr
-}
-
-func (ms *MemStore) Commit(name string, data []byte) error {
-	if ms.readonly {
-		return errReadOnly
-	}
-	d := ms.data[name]
-	d.data = data
-	return nil
-}
-
-func (ms *MemStore) Remove(name string) error {
-	delete(ms.data, name)
-	return nil
-}
-
-func (ms *MemStore) ReadOnly(ro bool) {
-	ms.readonly = ro
-}
-
-func (ms *MemStore) Disable(disable bool) {
-	ms.disable = disable
-}
-
-func NewMemStore() *MemStore {
-	return &MemStore{
-		data: make(map[string]*MemStoreData),
-	}
 }
 
 type MockStore struct {
@@ -161,13 +64,13 @@ func (ms *MockStore) Close() error {
 	return ret.Error(0)
 }
 
-func (ms *MockStore) OpenWrite(name string) (WriteCloserCommitter, error) {
+func (ms *MockStore) OpenWrite(name string) (utils.WriteCloserCommitter, error) {
 	ret := ms.Called(name)
 	wcc := ret.Get(0)
 	if wcc == nil {
 		return nil, ret.Error(1)
 	}
-	return ret.Get(0).(WriteCloserCommitter), ret.Error(1)
+	return ret.Get(0).(utils.WriteCloserCommitter), ret.Error(1)
 }
 
 func (ms *MockStore) OpenRead(name string) (io.ReadCloser, error) {
